+++ CppCore/gtests/TestConFrameHelpers.cc
@@ -11,7 +11,7 @@
 
 TEST(SymbolToAtomicNumberTest, ValidSymbols) {
   std::vector<std::string> symbols = {"H", "O", "N", "Fe", "Au"};
-  std::vector<size_t> expected = {1, 8, 7, 26, 79};
+  std::vector<int> expected = {1, 8, 7, 26, 79};
 
   EXPECT_EQ(yodecon::symbols_to_atomic_numbers(symbols), expected);
 }
@@ -31,27 +31,27 @@ TEST(SymbolToAtomicNumberTest, EmptyInput) {
 
 TEST(SymbolToAtomicNumberTest, SingleSymbol) {
   std::vector<std::string> symbols = {"O"};
-  std::vector<size_t> expected = {8};
+  std::vector<int> expected = {8};
 
   EXPECT_EQ(yodecon::symbols_to_atomic_numbers(symbols), expected);
 }
 
 TEST(AtomicNumberToSymbolTest, TestValidAtomicNumbers) {
-  std::vector<size_t> atomic_numbers = {1, 6, 8, 15};
+  std::vector<int> atomic_numbers = {1, 6, 8, 15};
   std::vector<std::string> expected_symbols = {"H", "C", "O", "P"};
   EXPECT_EQ(yodecon::atomic_numbers_to_symbols(atomic_numbers),
             expected_symbols);
 }
 
 TEST(AtomicNumberToSymbolTest, TestInvalidAtomicNumber) {
-  std::vector<size_t> atomic_numbers = {
-      1, 300}; // 300 is not a valid atomic number
+  std::vector<int> atomic_numbers = {1,
+                                     300}; // 300 is not a valid atomic number
   EXPECT_THROW(yodecon::atomic_numbers_to_symbols(atomic_numbers),
                std::invalid_argument);
 }
 
 TEST(AtomicNumberToSymbolTest, TestEmptyInput) {
-  std::vector<size_t> atomic_numbers = {};
+  std::vector<int> atomic_numbers = {};
   std::vector<std::string> expected_symbols = {};
   EXPECT_EQ(yodecon::atomic_numbers_to_symbols(atomic_numbers),
             expected_symbols);
